+++ trip planner/trip_plannerApp.swift
@@ -6,12 +6,20 @@
 //
 
 import SwiftUI
+import GoogleSignIn
 
 @main
 struct trip_plannerApp: App {
+    init() {
+        GoogleAuth.configure()
+    }
+
     var body: some Scene {
         WindowGroup {
             ContentView()
+                .onOpenURL { url in
+                    GIDSignIn.sharedInstance.handle(url)
+                }
         }
     }
 }